requirements.txt
requests~=2.27
mozilla-django-oidc==4.0.1
django-ninja
